requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.9
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
six==1.15.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
zipp==3.21.0
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
python-dotenv==0.19.0